requirements.txt
Flask
gunicorn
requests
Werkzeug
opencv-python
opencv-python-headless
betterocr
Pillow
Flask-SocketIO
pyheif
eventlet